Rename wp:theme-file-uris to wp:theme-file Rename utils file
